requirements.txt
flask==3.0.0
azure-storage-blob==12.19.0
flask_cors==4.0.0
openai==1.12.0
langchain==0.1.7
langchain-core==0.1.23
slate3k==0.5.3
python-dotenv==1.0.0
langchain-openai==0.0.6
google-generativeai==0.3.2
gtts==2.5.1
psycopg2==2.9.9
gunicorn==21.2.0